requirement.txt
mecab-python3
unidic-lite
jupyter
pandas
matplotlib
japanize-matplotlib
